Refactor status log to catch errors
